# lab5/openapi.yaml
openapi: 3.0.0
info:
  title: Conference Management API
  description: API for managing users, reviews and conferences
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: User service
  - url: http://localhost:8001
    description: Review service
  - url: http://localhost:8002
    description: Conference service
paths:
  /token:
    post:
      summary: Get JWT token
      tags: [auth]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
...
  /users/:
    post:
      summary: Create a new user
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        full_name:
          type: string
        email:
          type: string
        disabled:
          type: boolean
    UserCreate:
      type: object
      properties:
        username:
          type: string
        full_name:
          type: string
        email:
          type: string
      required:
        - username
...