# lab5/user_service/app/models.py
from .database import Base
from sqlalchemy import Column, Integer, String, Boolean

class User(Base):
    __tablename__ = "users"

    id = Column(Integer, primary_key=True, index=True)
    username = Column(String(50), unique=True, index=True)
    full_name = Column(String(100))
    email = Column(String(100))
    hashed_password = Column(String(100))
    disabled = Column(Boolean, default=False)